+++ routes/question-cerb-exhausted/schema.js
@@ -1,0 +1,14 @@
+/* istanbul ignore file */
+
+const Schema = {
+  cerb_exhausted: {
+    isIn: {
+      errorMessage: 'errors.multipleChoiceGeneric',
+      options: [['cerb-exhausted-yes', 'cerb-exhausted-almost', 'cerb-exhausted-no']],
+    },
+  },
+}
+
+module.exports = {
+  Schema,
+}