+++ routes/question-cerb/schema.js
@@ -1,0 +1,14 @@
+/* istanbul ignore file */
+
+const Schema = {
+  cerb: {
+    isIn: {
+      errorMessage: 'errors.multipleChoiceGeneric',
+      options: [['receiving-cerb', 'not-receiving-cerb']],
+    },
+  },
+}
+
+module.exports = {
+  Schema,
+}