+++ routes/question-province/js/question-province.js
@@ -4,6 +4,7 @@
 
 accessibleAutocomplete.enhanceSelectElement({
   selectElement: document.querySelector('#province-select'),
+  autoselect: false,
   showAllValues: true,
   showNoOptionsFound: false,
   dropdownArrow: () => {