+++ routes/question-your-situation-no-income/question-your-situation-no-income.controller.js
@@ -27,5 +27,5 @@ const postNoIncome = (req, res) => {
     'rrif',
     'gross_income',
   ])
-  return res.redirect(res.locals.routePath('question-mortgage-payments'))
+  return res.redirect(res.locals.routePath('question-cerb'))
 }