+++ routes/question-your-situation-some-income/question-your-situation-some-income.controller.js
@@ -33,7 +33,7 @@ const postSomeIncome = (req, res) => {
 
   if (['quarantine','none-of-the-above'].includes(req.body.some_income)) {
     pruneSessionData(req, ['rrif', 'reduced-income', 'gross-income'])
-    return res.redirect(res.locals.routePath('question-mortgage-payments'))
+    return res.redirect(res.locals.routePath('question-cerb'))
   }
 
   if (req.body.some_income === 'retired') {