+++ routes/question-your-situation-some-income/question-your-situation-some-income.controller.spec.js
@@ -30,7 +30,7 @@ describe('Test redirects for some-income ', () => {
       ],
     },
     {
-      dest: 'question-mortgage-payments',
+      dest: 'question-cerb',
       values: ['quarantine', 'none-of-the-above'],
     },
   ]